+++ docs/assets/chunks/wcpay.d0e1a780.js
@@ -1,0 +1,1 @@
+const s="/assets/alipay.4018db41.png",a="/assets/wcpay.44f75c57.png";export{s as _,a};